# storage/swagger/http.json
{"openapi":"3.0.0","info":{"title":"Tecno Pix API","description":"API REST para saque PIX imediato e agendado.","version":"0.1.0"},"servers":[{"url":"http:\/\/127.0.0.1:9501","description":"Tecno Pix API"}],"paths":{"\/accounts":{"get":{"tags":["Accounts"],"summary":"Lista as contas cadastradas","operationId":"listAccounts","responses":{"200":{"description":"Lista de contas","content":{"application\/json":{"schema":{"type":"array","items":{"$ref":"#\/components\/schemas\/Account"}}}}}}}},"\/":{"get":{"tags":["System"],"summary":"Endpoint inicial da API","operationId":"rootIndex","responses":{"200":{"description":"Resposta padrao da API","content":{"application\/json":{"schema":{"properties":{"method":{"type":"string","example":"GET","rules":null,"attribute":null},"message":{"type":"string","example":"Hello Hyperf.","rules":null,"attribute":null}},"type":"object"}}}}}}},"\/account\/{accountId}\/balance\/withdraw":{"post":{"tags":["Withdraw"],"summary":"Solicita um saque para uma conta","operationId":"createWithdraw","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application\/json":{"schema":{"required":["method","amount","pix"],"properties":{"method":{"type":"string","example":"pix","rules":null,"attribute":null},"amount":{"type":"number","format":"float","example":100.5,"rules":null,"attribute":null},"pix":{"properties":{"type":{"type":"string","example":"email","rules":null,"attribute":null},"key":{"type":"string","example":"cliente@example.com","rules":null,"attribute":null}},"type":"object","rules":null,"attribute":null},"schedule":{"type":"string","format":"date-time","nullable":true,"rules":null,"attribute":null}},"type":"object"}}},"_content":{"required":["method","amount","pix"],"properties":{"method":{"type":"string","example":"pix","rules":null,"attribute":null},"amount":{"type":"number","format":"float","example":100.5,"rules":null,"attribute":null},"pix":{"properties":{"type":{"type":"string","example":"email","rules":null,"attribute":null},"key":{"type":"string","example":"cliente@example.com","rules":null,"attribute":null}},"type":"object","rules":null,"attribute":null},"schedule":{"type":"string","format":"date-time","nullable":true,"rules":null,"attribute":null}},"type":"object"}},"responses":{"202":{"description":"Solicitacao aceita para processamento","content":{"application\/json":{"schema":{"properties":{"message":{"type":"string","example":"Withdraw request accepted.","rules":null,"attribute":null},"id":{"type":"string","format":"uuid","rules":null,"attribute":null},"account_id":{"type":"string","format":"uuid","rules":null,"attribute":null},"method":{"type":"string","example":"pix","rules":null,"attribute":null},"amount":{"type":"string","example":"100.50","rules":null,"attribute":null},"scheduled":{"type":"boolean","example":false,"rules":null,"attribute":null}},"type":"object"}}}},"422":{"description":"Payload ou regra de negocio invalida","content":{"application\/json":{"schema":{"properties":{"message":{"type":"string","example":"The given data was invalid.","rules":null,"attribute":null},"errors":{"type":"object","rules":null,"attribute":null}},"type":"object"}}}},"404":{"description":"Conta nao encontrada","content":{"application\/json":{"schema":{"properties":{"message":{"type":"string","example":"Account not found.","rules":null,"attribute":null}},"type":"object"}}}}}}}},"components":{"schemas":{"Account":{"required":["id","name","balance"],"properties":{"id":{"type":"string","format":"uuid","example":"8b61f3e5-85c6-4e8f-9e1d-9c13f5eaf001","rules":null,"attribute":null},"name":{"type":"string","example":"Conta Demo","rules":null,"attribute":null},"balance":{"type":"string","example":"1000.00","rules":null,"attribute":null},"created_at":{"type":"string","format":"date-time","rules":null,"attribute":null},"updated_at":{"type":"string","format":"date-time","rules":null,"attribute":null}},"type":"object"}}},"tags":[{"name":"Accounts","description":"Consultas de contas"},{"name":"System","description":"Rotas de sistema"},{"name":"Withdraw","description":"Solicitacao de saque"},{"name":"System","description":"Rotas de sistema"},{"name":"Accounts","description":"Rotas de consulta de contas"}]}